requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
distro==1.9.0
dnspython==1.16.0
fastapi==0.116.1
filelock==3.19.1
firebase_admin==7.1.0
frozenlist==1.7.0
future==1.0.0
gevent==25.8.2
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-genai==1.32.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-aiohttp==0.1.8
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
msgpack==1.1.1
multidict==6.6.4
openai==1.102.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==3.12.0
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.5
requests-file==2.1.0
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
starlette==0.47.3
tenacity==9.1.2
tldextract==5.3.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
yarl==1.20.1
zope.event==5.1.1
zope.interface==7.2
